requirements.txt
google-cloud-bigquery
google-cloud-storage
pandas
sqlalchemy
pyarrow
fastparquet
configparser
pymssql
pymysql
pyodbc
